use create-react-native-app
modify the demo to create-react-native-app
