--- ReactNative/SmartGarden/.buckconfig
@@ -1,6 +1,0 @@
-
-[android]
-  target = Google Inc.:Google APIs:23
-
-[maven_repositories]
-  central = https://repo1.maven.org/maven2
