--- ReactNative/SmartGarden/android/keystores/BUCK
@@ -1,8 +1,0 @@
-keystore(
-    name = "debug",
-    properties = "debug.keystore.properties",
-    store = "debug.keystore",
-    visibility = [
-        "PUBLIC",
-    ],
-)
