--- ReactNative/SmartGarden/android/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name = 'SmartGarden'
-
-include ':app'
