--- ReactNative/SmartGarden/app.json
@@ -1,4 +1,0 @@
-{
-  "name": "SmartGarden",
-  "displayName": "SmartGarden"
-}
