--- ReactNative/SmartGarden/index.js
@@ -1,4 +1,0 @@
-import { AppRegistry } from 'react-native';
-import App from './App';
-
-AppRegistry.registerComponent('SmartGarden', () => App);
