--- ReactNative/SmartGarden/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "SmartGarden",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
-  },
-  "dependencies": {
-    "react": "16.3.1",
-    "react-native": "0.55.4"
-  },
-  "devDependencies": {
-    "babel-jest": "23.2.0",
-    "babel-preset-react-native": "4.0.0",
-    "jest": "23.2.0",
-    "react-test-renderer": "16.3.1"
-  },
-  "jest": {
-    "preset": "react-native"
-  }
-}
