+++ ReactNative/demo/app.json
@@ -1,0 +1,5 @@
+{
+  "expo": {
+    "sdkVersion": "27.0.0"
+  }
+}
